Move CoachTest interface
