+++ src/app/interfaces/coach-edit.ts
@@ -1,0 +1,7 @@
+export interface CoachTest {
+  id: string;
+  level: string;
+  isChecked: boolean;
+  isHigh: boolean;
+  date: string;
+}
